examples/simple_math_sft/requirements.txt
# Simple Math SFT Example Dependencies
#
# Install with: pip install -r requirements.txt
# For TPU: pip install -r requirements.txt -f https://storage.googleapis.com/jax-releases/libtpu_releases.html

# JAX (CPU/GPU version - TPU version installed separately)
jax[cpu]>=0.4.30

# Flax NNX (modern neural network API)
flax>=0.8.0

# Bonsai - JAX model implementations
jax-bonsai @ git+https://github.com/jax-ml/bonsai.git

# Optax - JAX optimizer library
optax>=0.2.0

# Orbax - Checkpointing
orbax-checkpoint>=0.5.0

# Grain - Data loading
grain>=0.2.0

# HuggingFace Transformers (for tokenizer)
transformers>=4.40.0

# HuggingFace Datasets (for GSM8K)
datasets>=2.18.0

# Utilities
numpy>=1.24.0
